backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pillow==10.4.0
numpy==1.26.4
aiofiles==24.1.0
pydantic==2.9.0
torch==2.4.1
torchvision==0.19.1
ultralytics==8.2.100
